Turning off unfurling in Slack API call
Slack, by default, unfurls media and urls.  When I send a DM message
with chat.postMessage the URL is activated by slack as part of the
unfurling process.  I had to disable unfurling because the user will
need to complete enrollment by clicking the link.  Slack was completing
enrollment on behalf of the user during unfurling.